refactor : 미사용 메서드 제거, 로그 제거 #SCRUM-116
