+++ src/main/java/com/example/letmovie/domain/reservation/dto/request/ScreenDTO.java
@@ -1,4 +1,4 @@
-package com.example.letmovie.domain.reservation.dto;
+package com.example.letmovie.domain.reservation.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
