+++ src/main/java/com/example/letmovie/domain/reservation/entity/ReservationStatus.java
@@ -8,7 +8,7 @@ public enum ReservationStatus {
     PENDING("예매 중"),         // 예매 중
     CANCELLED("취소"),         // 취소
     COMPLETED("예매 완료"),     // 예매 완료
-    VIEWED("상영 확인");        // 상영 확인 todo 상영시간 확인 후 지나면 알아서 바뀌게 해야 함.
+    VIEWED("상영 확인");        // 상영 확인
 
     private final String displayName;
