+++ src/main/java/com/example/letmovie/domain/reservation/service/ReservationService.java
@@ -12,7 +12,6 @@ import com.example.letmovie.domain.reservation.entity.Reservation;
 import com.example.letmovie.domain.reservation.entity.ReservationSeat;
 import com.example.letmovie.domain.reservation.entity.ReservationStatus;
 import com.example.letmovie.domain.reservation.entity.Seat;
-import com.example.letmovie.domain.reservation.facade.OptimisticLockReservationFacade;
 import com.example.letmovie.domain.reservation.repository.ReservationRepository;
 import com.example.letmovie.domain.reservation.repository.SeatRepository;
 import com.example.letmovie.domain.reservation.repository.ShowtimeRepository;
@@ -96,14 +95,12 @@ public class ReservationService {
     public void reservationCancel(Long reservationId) {
             Reservation reservation = reservationRepository.findById(reservationId).orElseThrow(ReservationNotFound::new);
             Payment payment = paymentRepository.findByReservationId(reservationId).orElseThrow(() -> new PaymentException(ErrorCodes.PAYMENT_NOT_FOUND));
-            log.info("예매취소시작");
-            log.info("예매취소시작2");
             PaymentResponse.Cancel cancelResult = paymentService.cancel(payment.getId());
+
         // 결제 취소가 성공하면 예매 취소 진행
         if (cancelResult.status().equals("CANCEL_PAYMENT")) {
             reservation.cancelReservation();
         }
-        log.info("예매취소시작3");
 
     }
 }
